+++ src/Data/Tax/ATO/Common.hs
@@ -195,9 +195,10 @@ instance (Foldable t, HasTaxWithheld x a a, Num a)
 class HasCentsPerKilometreMethod y where
   centsPerKilometre :: Num a => a
 
-instance (2016 <= y, y <= 2026, FinancialYear y) => HasCentsPerKilometreMethod y where
+instance (2016 <= y, y <= 2027, FinancialYear y) => HasCentsPerKilometreMethod y where
   centsPerKilometre = case fromProxy (Proxy @y) of
-    y | y == 2026           -> 88
+    y | y == 2027           -> 88  -- TODO confirm
+      | y == 2026           -> 88
       | y == 2025           -> 88
       | y == 2024           -> 85
       | y == 2023           -> 78
@@ -227,12 +228,13 @@ applyCentsPerKilometreMethod kms = Money $ centsPerKilometre @y * min kms 5000 /
 class HasFixedRateMethod y where
   fixedRateMethodCentsPerHour :: Num a => a
 
-instance (2023 <= y, y <= 2026, FinancialYear y) => HasFixedRateMethod y where
+instance (2023 <= y, y <= 2027, FinancialYear y) => HasFixedRateMethod y where
   fixedRateMethodCentsPerHour = case fromProxy (Proxy @y) of
     y | y == 2023           -> 67
       | y == 2024           -> 67
       | y == 2025           -> 70
       | y == 2026           -> 70
+      | y == 2027           -> 70  -- TODO confirm
       | otherwise           ->  0  -- can't happen
 
 -- | The updated fixed rate method available from the 2022–23 income